Add PartenaireController implementation
